requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
scikit-learn
matplotlib
seaborn
requests
numpy
